+++ lib/sixarm_ruby_time_stamp/time.rb
@@ -5,7 +5,7 @@ Please see README
 
 class Time
 
-  # Create a time stamp string using the format "YYYY-MM-DDTHH:MM:SSZ"
+  # Create a time stamp string using the default format "YYYY-MM-DDTHH:MM:SSZ"
   #
   # This standard format is specified in IETF RFC 3339 and ISO 8601.
   #
@@ -19,8 +19,8 @@ class Time
   #
   # Return: [String] a time stamp string
   #
-  def stamp
-    getutc.strftime('%Y-%m-%dT%H:%M:%SZ')
+  def stamp(format = stamp_format)
+    getutc.strftime(format)
   end
 
   # Shorthand for Time.now.stamp
@@ -32,8 +32,66 @@ class Time
   #
   # Return: [String] Time.now time stamp string
   #
-  def self.stamp
-    now.stamp
+  def self.stamp(format = self.stamp_format)
+    now.stamp(format)
+  end
+
+  # Get stamp format string
+  #
+  # Example:
+  #
+  #     time.stamp_format => "%Y-%m-%dT%H:%M:%S.%NZ"
+  #
+  # Return: [String] stamp format string
+  #
+  def stamp_format
+    @stamp_format ||= '%Y-%m-%dT%H:%M:%S.%NZ'
+  end
+
+  # Set stamp format string
+  #
+  # Example:
+  #
+  #     time.stamp_format = "%Y-%m-%dT%H:%M:%S.%NZ"
+  #
+  def stamp_format=(format)
+    @stamp_format = format
+  end
+
+  # Get class-scope stamp format string
+  #
+  # Example:
+  #
+  #     Time.stamp_format => "%Y-%m-%dT%H:%M:%S.%NZ"
+  #
+  # Return: [String] stamp format string
+  #
+  def self.stamp_format
+    @@stamp_format ||= '%Y-%m-%dT%H:%M:%S.%NZ'
+  end
+
+  # Set class-scope stamp format string
+  #
+  # Example:
+  #
+  #     Time.stamp_format = "%Y-%m-%dT%H:%M:%S.%NZ"
+  #
+  # Return: [String] stamp format string
+  #
+  def self.stamp_format=(format)
+    @@stamp_format = format
+  end
+
+  # Get class-scope stamp format default string
+  #
+  # Example:
+  #
+  #     Time.stamp_format_default => "%Y-%m-%dT%H:%M:%S.%NZ"
+  #
+  # Return: [String] stamp format default string
+  #
+  def self.stamp_format_default
+    "%Y-%m-%dT%H:%M:%S.%NZ"
   end
 
 end